feat: style adjustments
Major overhaul of styling. Specifically, customize more elements
and apply circuit-ui colors in more places. Overall, gives the website
nicer feel and distinguishes it from the default Starlight template.

Still not sure if it's worth the additional code that needs to be
maintained, so we might eventual decide to ease up on some of the
customization and go back to default. On the other hand, some
customization is necessary because we for example rely heavily on top
navigation to segment the developer portal into a few main sections
(online payments, terminal payments, API reference) which is not
supported by default.